// src/types/GogoAnime/AnimeSearch.ts
// Generated by https://quicktype.io
//
// To change quicktype's target language, run command:
//
//   "Set quicktype target language"

export default interface Search {
    currentPage: string;
    hasNextPage: boolean;
    results:     Result[];
}

export interface Result {
    id:          string;
    title:       string;
    url:         string;
    image:       string;
    releaseDate: string;
    subOrDub:    SubOrDub;
}

export enum SubOrDub {
    Dub = "dub",
    Sub = "sub",
}